@@ -8,22 +8,22 @@
   "dependencies": {
     "ms": "^2.1.3",
     "mysql": "^2.18.1",
-    "mysql2": "^3.5.2",
+    "mysql2": "^3.11.0",
     "next": "latest",
     "objects-to-csv": "^1.3.6",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "sequelize": "^6.32.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "sequelize": "^6.37.3",
     "serverless-mysql": "^1.5.5",
-    "ssh2": "^1.14.0",
-    "swr": "^2.0.0",
+    "ssh2": "^1.15.0",
+    "swr": "^2.2.5",
     "uid": "^2.0.2"
   },
   "devDependencies": {
     "@types/mysql2": "github:types/mysql2#semver:1.0.0",
-    "@types/node": "^18.0.0",
-    "@types/react": "^18.0.14",
-    "@types/react-dom": "^18.0.5",
-    "typescript": "^4.7.4"
+    "@types/node": "^18.19.50",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "typescript": "^4.9.5"
   }
 }
